requirements.txt
streamlit
pandas
scikit-learn
azure-storage-blob